# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/bills": {
            "get": {
                "description": "Get bills (paginated) — id, type, storeId, storeName, bookNo,\nreceiptNo, customerName, customerAddress, total, itemCount, hasSlip, createdAt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bill Module (Version 1)"
                ],
                "summary": "List bills",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search keyword",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by bill type: receipt or payment",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "filter by store id",
                        "name": "storeId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter created_at \u003e= this date, YYYY-MM-DD",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter created_at \u003c= this date, YYYY-MM-DD",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/paginator.Pagination"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a bill (JSON) with one or more line items, priced by the\ncaller (no price list). Book/receipt numbers are computed\nserver-side, scoped per (store, type); unit/subtotal/total and\nanything else derivable are ignored if the client sends them.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bill Module (Version 1)"
                ],
                "summary": "Create bill",
                "parameters": [
                    {
                        "description": "bill",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateBill"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.BillDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/bills/{id}": {
            "get": {
                "description": "Get bill by id — every bills column except deleted_at, plus a ready-to-use slipUrl",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bill Module (Version 1)"
                ],
                "summary": "Get bill",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "bill id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.BillDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Replace a bill's customer, items, discount, and shipping fee\n(JSON, same shape as create). storeId and type must match the\nexisting bill — sending different values is rejected with 400.\nBook/receipt numbers and the slip are never touched here.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bill Module (Version 1)"
                ],
                "summary": "Update bill",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "bill id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "bill",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateBill"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.BillDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete bill by id — soft delete only, slip and book/receipt number are kept as-is (never reused)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bill Module (Version 1)"
                ],
                "summary": "Delete bill",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "bill id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http_response.OkResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/bills/{id}/slip": {
            "put": {
                "description": "Upload a slip image (jpeg/png/webp, checked by magic bytes, ≤ 10MB) for an existing, non-deleted bill. Replaces any existing slip.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bill Module (Version 1)"
                ],
                "summary": "Upload a bill's slip",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "bill id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "slip image",
                        "name": "slip",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "slipUrl": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Clear a bill's slip and remove the underlying file (best-effort)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bill Module (Version 1)"
                ],
                "summary": "Delete a bill's slip",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "bill id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http_response.OkResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exception.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.CreateBill": {
            "type": "object",
            "properties": {
                "customerAddress": {
                    "type": "string"
                },
                "customerId": {
                    "type": "integer"
                },
                "customerName": {
                    "type": "string"
                },
                "customerPhone": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CreateBillItem"
                    }
                },
                "shippingFee": {
                    "type": "number"
                },
                "storeId": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dtos.CreateBillItem": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "productId": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                }
            }
        },
...
        "dtos.UpdateBill": {
            "type": "object",
            "properties": {
                "customerAddress": {
                    "type": "string"
                },
                "customerId": {
                    "type": "integer"
                },
                "customerName": {
                    "type": "string"
                },
                "customerPhone": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CreateBillItem"
                    }
                },
                "shippingFee": {
                    "type": "number"
                },
                "storeId": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "exception.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/exception.ParameterError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "exception.ParameterError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "http_response.OkResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "paginator.Pagination": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "object",
                    "additionalProperties": true
                },
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "orderBy": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "sort": {
                    "type": "string"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        },
        "responses.BillDetailResponse": {
            "type": "object",
            "properties": {
                "bookNo": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "customerAddress": {
                    "type": "string"
                },
                "customerId": {
                    "type": "integer"
                },
                "customerName": {
                    "type": "string"
                },
                "customerPhone": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.BillItemDetail"
                    }
                },
                "receiptNo": {
                    "type": "integer"
                },
                "shippingFee": {
                    "type": "number"
                },
                "slipUrl": {
                    "type": "string"
                },
                "storeId": {
                    "type": "integer"
                },
                "storeLogo": {
                    "type": "string"
                },
                "storeName": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "responses.BillItemDetail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "productId": {
                    "type": "integer"
                },
                "productName": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "subtotal": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
    }
}